Move AlignmentErrors to a new error package
