+++ render-ws-java-client/src/main/java/org/janelia/render/client/newsolver/StackAlignmentComparisonClient.java
@@ -3,7 +3,8 @@ package org.janelia.render.client.newsolver;
 import com.beust.jcommander.Parameter;
 import org.janelia.alignment.match.OrderedCanvasIdPairWithValue;
 import org.janelia.render.client.ClientRunner;
-import org.janelia.render.client.newsolver.AlignmentErrors.MergingMethod;
+import org.janelia.render.client.newsolver.errors.AlignmentErrors;
+import org.janelia.render.client.newsolver.errors.AlignmentErrors.MergingMethod;
 import org.janelia.render.client.parameter.CommandLineParameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
