+++ render-ws-java-client/src/main/java/org/janelia/render/client/newsolver/errors/AlignmentErrors.java
@@ -1,4 +1,4 @@
-package org.janelia.render.client.newsolver;
+package org.janelia.render.client.newsolver.errors;
 
 import org.janelia.alignment.match.OrderedCanvasIdPair;
 import org.janelia.alignment.match.OrderedCanvasIdPairWithValue;
@@ -12,7 +12,7 @@ import java.util.Map;
 import java.util.function.DoubleBinaryOperator;
 import java.util.stream.Collectors;
 
-class AlignmentErrors {
+public class AlignmentErrors {
 	private final List<OrderedCanvasIdPairWithValue> pairwiseErrors = new ArrayList<>();
 
 	public void addError(final OrderedCanvasIdPair pair, final double error) {
